feat: add "Edit" button
